Split Mesh and Material into separate structs
This makes multiple objects able to work on the same meshes/materials,
thus saving us from having to create more than one of each.
