+++ src/scene.rs
@@ -1,10 +1,9 @@
-use ash::vk;
+use ash::{vk, Device};
 use cgmath::prelude::*;
 use cgmath::{Deg, Matrix4, Point3, Quaternion, Vector3};
 use object::transform::Transformable;
-use object::{DrawObject, Drawable};
+use object::{DrawObject, Drawable, Material, Mesh};
 use renderer::{MainPass, RenderState};
-use std::f32;
 
 pub struct Scene
 {
@@ -19,7 +18,21 @@ impl Scene
 			objects: Vec::new(),
 		};
 
-		let mut cuboid = DrawObject::new_cuboid(rs, mp, Point3::new(0.0, 0.0, -4.0), 2.0, 2.0, 2.0);
+		let quad_mesh = Mesh::new_quad(rs, 20.0, 20.0);
+		let cuboid_mesh = Mesh::new_cuboid(rs, 2.0, 2.0, 2.0);
+		let metal_panel_surface = Material::new(
+			rs,
+			mp,
+			"assets/thirdparty/textures/Metal_Panel_004/Metal_Panel_004_COLOR.jpg",
+			"assets/thirdparty/textures/Metal_Panel_004/Metal_Panel_004_NORM.jpg",
+		);
+		let cube_surface = Material::new(
+			rs,
+			mp,
+			"assets/original/textures/cubemap.png",
+			"assets/original/textures/cubemap_normals.png",
+		);
+		let cuboid = DrawObject::new(Point3::new(0.0, 0.0, -4.0), cuboid_mesh, cube_surface);
 		scene.objects.push(cuboid);
 
 		let points = vec![
@@ -45,7 +58,7 @@ impl Scene
 			let x: f32 = points[i].x;
 			let y: f32 = points[i].y;
 			let z: f32 = points[i].z;
-			let mut wall = DrawObject::new_quad(rs, mp, Point3::new(0., 0., 0.), 20.0, 20.0);
+			let mut wall = DrawObject::new(Point3::new(0., 0., 0.), quad_mesh.clone(), metal_panel_surface.clone());
 			wall.globally_rotate(Quaternion::from_axis_angle(directions[i], Deg(90.0)));
 			if i == 5
 			{
@@ -73,13 +86,13 @@ impl Scene
 	}
 
 	pub fn draw(
-		&self, cmd_buf: vk::CommandBuffer, pipeline_layout: vk::PipelineLayout, view_matrix: &Matrix4<f32>,
-		projection_matrix: &Matrix4<f32>,
+		&self, device: &Device, cmd_buf: vk::CommandBuffer, pipeline_layout: vk::PipelineLayout,
+		view_matrix: &Matrix4<f32>, projection_matrix: &Matrix4<f32>,
 	)
 	{
 		for object in self.objects.iter()
 		{
-			object.draw(cmd_buf, pipeline_layout, view_matrix, projection_matrix);
+			object.draw(device, cmd_buf, pipeline_layout, view_matrix, projection_matrix);
 		}
 	}
 }
